finnie_ai/requirements.txt
# Finnie AI - Requirements
# Python 3.12

# LLM + Agents
langchain==0.2.17
langchain-community==0.2.17
langchain-openai==0.1.23
langgraph==0.2.45
openai==1.43.0
langchain-chroma==0.1.4

# RAG / Vector DB
chromadb==0.5.3

# Data
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4

# Market Data APIs
yfinance==0.2.40
alpha_vantage==2.3.1
finnhub-python==2.4.18
polygon-api-client==1.14.0

# Search APIs
tavily-python==0.3.3
google-search-results==2.4.2

# UI
streamlit==1.37.1

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic==2.8.2
typing-extensions==4.12.2
httpx==0.27.0